+++ microservices/flight-service/src/main/java/com/esteban/flightservice/service/FlightService.java
@@ -19,5 +19,5 @@ public interface FlightService {
 
 	List<Flight> getFlightsByPurchaseId(Long purchaseId);
 	
-	Double setPricePerUser (int basePrice, int age, boolean luggage);
+	Double setPricePerUser (Double basePrice, int age, boolean luggage);
 }
